Explicitly specify the entry point WinMainCRTStartup.
This produces a program that runs on Windows too. Note that I am calling
gcc, not ld, to link this. With ld it is not including some static
objects from /usr/x86_64-w64-mingw32/lib that contain necessary code, in
particular WinMainCRTStartup.

Let's see if that works for André too.