Don't export SHCLSIDFromString by ordinal for now.
Mingw creates an entry in the export table that is not sorted correctly
if we do this, leading to the PE loader loading the wrong import. I
guess we'll need something similar to Wine's spec files that generates
ordinals for export by name only functions and works around mingw's
quirkyness.