kernel32: Use callback_helper.h
It isn't exactly the same, but keeping the functions around is probably
better anyway because we otherwise alloc and free a lot. The constant
entries are also needed for another function (I forgot the name), which
doesn't have a clear point to free the wrapper (It would die with the
handle, which would make CloseHandle really complicated).