Drop the SetLastError hook.
It doesn't work on Macbecause the Wine headers inline SetLastError. For
some reason the inlining didn't work on aarch64 Linux, making the hooking
trick work.

No benchmarking has been done on this yet. It may be an idea later on
to build Wine with SetLastError inlining disabled and removing the constant
syncing.