advapi32: Verify a few PSID related functions.
So far I am hoping that applications don't look inside the PSID pointer.
It seems to be a struct that is defined deep in the NT internals, so
presumably applications have no business peeking there. If they do
depend on the size and contents I have a mess on my hands.