Remove a lot of DLL wrappers.
All these DLLs have a massive interface to the application, but are
buildable as PE DLLs. In particular shell32 is a mess - see the commit
sequence leading up to this patch, and these commits only touched a
small part of shell32.

To make matters worse, shell32 drags in OLE32, so we either have the
choice of living with two OLE32 instances (bad) or thunking ole32 as
well (also bad).

This patch presumably gives up on running hangover on Windows.