ntdll: Don't forward RtlAddVectoredExceptionHandler.
It won't do the right thing and needs to be handled on the guest side.

This is called by binaries built with a pthread-enabled mingw, which I
happen to have because once upon a time I built dxvk on my Mac.
Forwarding this leads to a crash on the first benign exception, e.g.
inside OutputDebugString.